[PATCH] m68k: driver updates [6/20]
- kdev_t fixes
- bio fixes
- other cleanups and compile fixes

+++ drivers/scsi/amiga7xx.c
@@ -31,7 +31,7 @@
 #include<linux/stat.h>
 
 extern int ncr53c7xx_init (Scsi_Host_Template *tpnt, int board, int chip, 
-			   u32 base, int io_port, int irq, int dma,
+			   unsigned long base, int io_port, int irq, int dma,
 			   long long options, int clock);
 
 int __init amiga7xx_detect(Scsi_Host_Template *tpnt)
